refactored: inventory dtos is moved to inventory package -set input and update dto to final -added validation to input and update
